+++ test/CustomStateSet.test.ts
@@ -1,6 +1,32 @@
-import { fixture, html, expect, fixtureCleanup } from '@open-wc/testing';
+import '../dist/index.js';
+
+import { fixture, html, expect, fixtureCleanup, aTimeout } from '@open-wc/testing';
 import { ICustomElement } from '../dist';
 import { CustomStateSet } from '../src/CustomStateSet';
+
+class CustomElementAddStateInConstructor extends HTMLElement {
+  internals = this.attachInternals();
+  constructor() {
+    super();
+
+    this.internals.states.add('--foo');
+  }
+}
+
+const addStateTagName = 'custom-element-add-state-in-constructor';
+customElements.define(addStateTagName, CustomElementAddStateInConstructor);
+
+class CustomElementDeleteStateInConstructor extends HTMLElement {
+  internals = this.attachInternals();
+  constructor() {
+    super();
+
+    this.internals.states.delete('--foo');
+  }
+}
+
+const deleteStateTagName = 'custom-element-delete-state-in-constructor';
+customElements.define(deleteStateTagName, CustomElementDeleteStateInConstructor);
 
 describe('CustomStateSet polyfill', () => {
   let el: HTMLElement;
@@ -54,6 +80,34 @@ describe('CustomStateSet polyfill', () => {
     if (el.part) {
       expect(el.part.contains('state--foo')).to.be.false;
       expect(el.part.contains('state--bar')).to.be.false;
+    }
+  });
+
+  it('will use a timeout if a state is added in a constructor', async () => {
+    let el;
+    expect(() => {
+      el = document.createElement(addStateTagName);
+    }).not.to.throw();
+
+    if (window.CustomStateSet.isPolyfilled) {
+      await aTimeout(100);
+      expect(el.matches('[state--foo]')).to.be.true;
+    } else {
+      expect(el.matches(`:--foo`)).to.be.true;
+    }
+  });
+
+  it('will use a timeout if a state is deleted in a constructor', async () => {
+    let el;
+    expect(() => {
+      el = document.createElement(deleteStateTagName);
+    }).not.to.throw();
+
+    if (window.CustomStateSet.isPolyfilled) {
+      await aTimeout(100);
+      expect(el.matches('[state--foo]')).to.be.false;
+    } else {
+      expect(el.matches(`:--foo`)).to.be.false;
     }
   });
 });